Reorder metrics dashboard
This commit reorders the metrics dashboard to hopefully improve usability.
I've tried to group related metrics together and bumped up the more useful
ones to the top.

For 24 hours i've gone with:

row 1

- new conversations
- questions (count)

row 2

- topics
- answer with error status

row 3

- unanswerable questions
- completeness

row 4

- Guardrail triggers
- Answer guardrails failures

row 5

- Question routing guardrails failures
- Question routing labels

row 6

- answer feedback

I've moved answer feedback down to the bottom as there's no mechanism to
provide it via the API so it only shows web chat users which we're not
worried about during the experiment. I'm going to remove it in a subsequent
commit.

Other than that i've tried to group the most related things on the same
row or as close as possible. i.e grouping guardrails together, and putting
question routing labels next to the question routing guardrails failures.

I've bumped up answer with error status because it doesn't naturally fit
into any of the groups but is a useful metric to monitor. If we see a spike
in that it could indicate a problem with the service. Although hopefully
our other monitoring would pick that up first.