Use secure cookies in hosted environments
One of the outcomes of the ITHC report was that we could improve the
security of our cookies by making them secure (only sent over HTTPS).

This commit updates the session cookie and the conversation_id cookie to
be secure when in hosted environments (prod, staging, int and on Heroku).

The reason i'm doing all hosted envs is that Rails.env.production? returns
true on staging and int. If we want to only have it in prod. I can:

- add an env var to helm for staging and int
- add an env var to the various heroku environments to toggle it off

I don't see a massive downside to using secure cookies in all of the above though.

I had a go at testing this, but everything that I came up with felt so
contrived that it seemed pointless. I've decided to leave this untested,
but if anyone's got any ideas then please do let me know.