Create error_timeout answers in the GET answer API endpoint
We have the concept of an answer timing out in the Chat UI. When an answer
takes longer than 2 minutes to generate, we create an "error_timeout" answer
so that the user is able to ask another question.

We didn't add this to the API endpoint, so this commit adds that functionality.