Add non_llm_answer to AnswerStrategy enum
This strategy has been added to a branch used for load testing. When
that branch is deployed and answers are generated, we persist the answer
with the non_llm_answer strategy.

When we redeploy out main, it returns nil since the value isn't in our
enum. This commit adds the value to the enum so that we can read
those records and the question show page in the admin UI doesn't blow
up when it tries to humanise nil.