Update exception messages from healthchecks
Our healthchecks are accessible on the public internet and currently
we're returning messages from them which expose certain details about
our application and infrastructure.

While these messages are useful to us for debugging, it's probably not
wise to expose them publicly.

Platform Engineering have confirmed that there's no precedent for these
routes not to be public, so it seems the best option is to remove
potentially sensitive messages instead.