Set allowed hosts config value for production
Restricts the allowed hosts to prevent DNS rebinding attacks.

For example, if a client makes a request to
https://chat.publishing.service.gov.uk with a header of
`X-Forwarded-Host: evil.com` then they'll get a 401 response.

If the client requests https://chat.publishing.service.gov.uk with
`X-Forwarded-Host: chat.publishing.service.gov.uk` then the request will
be processed as normal.

If the `X-Forwarded-Host` header isn't set then this config is ignored
and the request is processed as normal.

So if the header is set, the value of the header must satisfy at least
one element in the hosts array.

The regex for Heroku apps is quite lenient as we need to make sure that
both regular (manually created) apps as well as review apps are allowed.
Because all of the Heroku apps for chat are under the same project, they
all start with `govuk-chat`.

We'll skip this check for any path beginning with "/healthcheck" as
Kubernetes makes these requests using an IP rather than a hostname, so
they all get blocked and the Kubernetes will kill the pod.

It's a bit tricky to test production config values in RSpec, so it feels
like the best way to ensure we don't mess up the regexes is to extract
the config into a module and test that independently.