Add create_paginatable_filter to test to reduce duplication
This commit extracts a bunch of similar code in the pagiation specs
into a h create_paginatable_filter that can be used to create filter
params.

I've updated the test to use the filters attributes to match the
expected params. This should mean that when we add additional filters
via attrs and don't add them to the pagination params, the tests will fail.