Rescue from URI::InvalidComponentError in LinkTokenMapper
We had an issue recently where some GOV.UK content contained a malformed
mailto URI. This caused an exception to be raised in the
LinkTokenMapper.

We should rescue these exceptions just like we do for
`URI::InvalidURIError` and store the malformed URI in the mapper as it
is.

There's not much we can do for URLs that are incorrect in the GOV.UK
content itself. We could try to fix this specific issue - the link that
triggered this exception was "mailto://consular.uae@fco.gov.uk" - but we
run the risk of changing content that we're not the owners of. It's best
to just keep the content as it is.