Populate forbidden_terms_detected when forbidden terms are found
This updates the ForbiddenTermsChecker to populate the new
forbidden_terms_detected attribute on the Answer model when forbidden
terms are found in the answer message.

I've extended to the comment to explain what scan returns and why we
use the second element of each sub-array.

The reason for this is that we return 3 capture groups in the regex

1. the preceding character before the forbidden term
2. the forbidden term itself
3. the following character after the forbidden term

Due to this we know we can always get the forbidden term from the second
element of each sub-array.