Fix bug with multiple consecutive forbidden terms
With the current implementation, when a match occurs the character after
the match is consumed as part of the match. This means that if there
are multiple consecutive forbidden terms, only the first is detected.

This updates the regex to use lookbehind and lookahead instead. This is
less performant than the orginal implementation since it has to backtrack
to check the lookbehind. It has increased the latency from ~0.008 seconds
to between ~0.01 - 0.03 seconds in my testing. Practically speaking this
is still very fast and the performance impact should be negligible.