Add link token mapping to LLM response
We want to add the mapping of link tokens to their corresponding URLs in
the LLM response that gets saved with the answer. This will give us
additional information about which links are being rendered in the answer
to the user. And give us additional insight into answers that trigger the
forbidden terms guardrail.