Default to grouping by created_at field
We have a lot of repetition of this grouping so we can simplify it by
making it a default argument.

I have made it a keyword argument as I intend in my next commit to add
another keyword argument, for counting a particular field, and I want
their purpose to be distinct.