Add an endpoint for distinct API end user metrics
This provides the means for a metric that will show counts of individual
end users who have used the API to ask at least 1 question in hour or
daily groupings.

The motivation for this is that it can be quite hard at the moment based
on the number of questions to know how many end users these originated
from and I have to keep doing stuff in the Rails console to check it.