Include number of end users in Slack summary
This enhances the information we print in Slack for the morning summary
to include the number of end users who asked questions.

In applying this I did a refactor of the question_summary method to have
a guard clause to reduce the amount of code spent in conditionals.

I also noticed that the approach we currently take for pulling out
question_summary makes use of the Answer table rather than Question as
the main source. This seems incorrect as answer timestamps and
question timestamps can vary. I've refactored the specs slightly to
accommodate both, with an intention to fix this issue in a subsequent
commit.