Use Question as the main data source for daily slack messages
This corrects some potentially inconsistent data in the daily slack
message by changing the primary data source to being the questions table.

By using questions table we:

- include question that have a pending status - which are an indication
  of a stuck question
- match the data in the admin area for filters, as answers are created
  after questions anything around midnight could end up slightly off

I've added a helper method to Question similar to the
group_by_aggregate_status to handle the case statement for a pending
status.

I also edited the sort_by for the list_items to sort alphabetically on
status should the values be the same. I'm not sure if this happened
before implicitly, but felt best to make it explicit.