Add sidekiq concurrency env var for Heroku review apps
A while ago, we added separate processes for answer generation and other
async tasks in Sidekiq. However, we didn't set the concurrency for Sidekiq
in our Heroku review apps. This means that the default concurrency of 10 is used,
which exceeds the db conns we have available on Heroku's basic plan.

This commit adds a SIDEKIQ_CONCURRENCY setting to app.json, setting it to 5,
which will keep the number of db connections within the basic plan limits.