Add reconnection functionality
Action cable automatically attempts to reconnect to a channel when it
detects a disconnection. However, the answer will have been fully
broadcast to the client by this point, atleast in this spike scenario.

This adds some additional functionality in. On the backend it:

- adds a answer method to the chat channel that
  - checks that there is an answer for the question
  - figures out where the disconnection took place based on the html param
    it recieves
  - broadcats the remainder of the answer to the client via the chat channel

Realistically we may just want to send the whole answer back and render
that.

In the client it:

- adds a reconnecting message to the chat conversation when a disconnection
  is detected
- removes the reconnecting message when the connection is re-established
- calls the answer method on reconnection to ensure the answer is rebroadcast
  from the point of disconnection