Use persistent websocket subscription for conversation
I was previously creating a new subscription for each question asked in a
conversation. My thought process was that it would probably be less expensive,
but after doing some research I found that ActionCable is designed to handle many
subscriptions efficiently, but creating and tearing down subscriptions frequently can
be more resource-intensive than maintaining a persistent connection.

This:

- updates the channel to broadcast using only a conversation id
- updates the job/answer method to broadcast to that channel
- passes the querstionId in the broadcast so the client can determine if
  the message is relevant to their current question
- updates the client to maintain a persistent subscription for the conversation
- updates the reconnect logic to re-request any missed answer content
  for the current question
- nulls the questionId when an answer has finished streaming or streaming
  has been stopped by the user
- removes the unsubscribe logic as it's no longer needed