Abort when LLM returns empty sources during answer composition
Previously we made a product decision that if the LLM did not provide any
sources as used we considered a question to be answered so long as the
answered property was true. As we didn't have means to know which sources
were used it was decided we should mark all as used.

In usage we found that this actually produced a peculiar result (a user
asked "how many t's are in universal credit?" and we gave an answer that
didn't use GOV.UK content). We found the evidence of all sources being
flagged as used to be misleading.

We have therefore made a new product decision that supersedes this
previous decision that if no known sources are used this is an indication
that there is insufficient evidence that the answer is grounded in GOV.UK
content and thus should be treated as if the "answered" flag was set to
false.