Abort when no sources are used after link token mapping
We call the Context#update_sources_from_exact_urls_used method after we
have checked that the LLM returned some used sources. However, this could
cause an issue if the following happens:

1. the llm returns sources_used with link tokens
2. after mapping the link tokens to exact_urls, none of the sources are
   actually used
3. we end up with an answer that has no used sources, but we never
   aborted the pipeline because we only checked the LLM response
   before mapping the link tokens to exact_urls.

To fix this, I've updated both the OpenAI and Claude structured answer
composers to abort no sources are used after mapping the link tokens to
exact_urls.

This actually caught a bug in the OpenAI system specs where the used source
was a base path rather than link_1.