Store cancelled answer message when stream in cancelled
When a user stops the streaming of an answer we need a way to ensure we
don't pick up additional content streamed for that answer.

Kev and I discusseed potentially killing the sidekiq job. That is definitely
an option, but i do wonder if it's better to just stick with what i've got.

Currently, I just set the questionId to null in the JS when the user cancels.
This means any further streamed content is ignored client side since i filter
the data against the questionId.

The advantage of this approach is that we can still store the full answer
for analyisis later.

This change adds a couple of new columns to the answers table to store
the partial cancelled message and a boolean to indicate the answer was cancelled.

When the user cancels the streaming, we make a call to the ChatChannel
to store the cancelled message and set the cancelled flag to true.

If we decide we want to stop the broadcast, we could send the job id in
the broadcasts and the client could pass that to the cancel action, which
could then use Sidekiq to find and stop the job.

Either option would work.