Alternate implementation where you cancel from within the job
Cancelling a sidekiq job once it's started running isn't as simply as
killing it from outside unless it's an iterable job. You have to kill the
worker.

There's some documentation on how to do this here:
https://github.com/sidekiq/sidekiq/wiki/FAQ#how-do-i-cancel-a-sidekiq-job

I went this this approach.

I'm not parcicularly convinced we need to do this though. My preference
would be to just ignore broadcasts that aren't related to the current
questionId on the front end.