Use postgres adapter
This reconfigures Action Cable to use PostgreSQL as its adapter instead of Redis
to see how easy it is to swap out.

There are some limitations though of using postgres' NOTIFY/LISTEN functionality

```
The “payload” string to be communicated along with the notification.
This must be specified as a simple string literal. In the default
configuration it must be shorter than 8000 bytes. (If binary data or
large amounts of information need to be communicated, it's best to
put it in a database table and send the key of the record.)
```
https://www.postgresql.org/docs/current/sql-notify.html

I think that should be fine because we're just sending tokens.

An article about performance issues
https://www.recall.ai/blog/postgres-listen-notify-does-not-scale

This bit is particularly gross

```
When a NOTIFY query is issued during a transaction, it acquires a global lock on the entire database (ref) during the commit phase of the transaction, effectively serializing all commits.
```