Get solid cable adapter working
This just spikes out how easy it is to get solid cable working as an
adapter for Action Cable.

It's also reasonably straight forward as you can see. I just followed
the guidance on the repo:
https://github.com/rails/solid_cable?tab=readme-ov-file#installation

The readme says

```
Solid Cable is a database-backed Action Cable adapter that keeps messages
in a table and continuously polls for updates. This makes it possible to
drop the common dependency on Redis, if it isn't needed for any other purpose.
Despite polling, the performance of Solid Cable is comparable to Redis
in most situations. And in all circumstances, it makes it easier to
deploy Rails when Redis is no longer a required dependency for Action
Cable functionality.
```