Chunking: include travel advice alert status
The `travel_advice` content schema includes an field within the
`details` field called `alert_status`[1]. This is a required field and
is an array of strings which represent any alerts about travelling to
the country.

The items in this array are labels representing the alert, e.g.
"avoid_all_but_essential_travel_to_parts". The array can also be empty
if there are no current alerts.

Currently we do not include this information when ingesting a
`travel_advice` content item. If there's an alert for a given country
then this is important information that we should be indexing,
retrieving and passing along to the LLM when a user asks a question
about travel to a country.

This commit adds a new `TravelGuideParser` class which will be used to
chunk `travel_advice` content types only.

It subclasses the `PartsContentParser` because it uses most of the
logic, but adds two new things.

First we overwrite the content item `description` field to suffix the
warnings if there are any. We use the description field when generating
the embeddings so it's an important part of every single document in the
OpenSearch cluster. We make a copy of the original `content_item` hash
in the initializer of this class to make sure we don't mutate the
original hash, in case it's used elsewhere.

Secondly we create an additional chunk for `travel_advice` documents if
there are any warnings. This chunk has a title of "Alert status" and
some HTML content that's a list of the warnings in a human-readable
format. When chunking the document, we add this as the first chunk. The
rest of the chunking is handled by the `PartsContentParser` as it was
previously.

[1]: https://github.com/alphagov/publishing-api/blob/main/content_schemas/formats/travel_advice.jsonnet#L31-L36