Update evaluation:generate_answer to take input ENV var
This updates the evaluation:generate_answer rake task to use the INPUT
env var to get the questions message rather than the QUESTION env var.

This brings it in line with the other evaluation rake tasks and will
allow the bulk evaluation rake task we're adding to have a shared interface.

I've also made a few small changes to the specs to algin them with the
other rake task specs. Such as:

- Using the input env var shared example
- moved the input into a let to avoid duplication