Update evaluation:generate_answer to return serialised answer
We're looking to retire the evaluation:generate_report task in favour of
the new evaluation:batch_process task. As part of this, we want to
ensure that the output of evaluation:generate_answer contains the data that
evaluation:generate_report provides to ensure the Data Scientists have
all the information they need.

This change updates the evaluation:generate_answer task to return the
serialised answer. This provides:

- all the answer data
- the sources used to generate the answer

We will lose the answer_strategy field in the output, but since this will
default to Claude which is the expected behaviour, and the user will
have to explicitly provide other providers, it shouldn't be needed.