Remove report_generator & generate_report rake task
Now we've updated the answer_generation rake task to return
Answer#serialize_for_evaluation instead of just the answer message
we can remove the report_generator and the generate_report rake task.

The batch_process rake task has been merged so we can use that in
conjunction with answer_generation to generate the outputs we need
for evaluation.