Add AnswerRelevancy metric and basic rake task
As mentioned in the previous PR we're spiking out adding a deepeval metric
to this codebase to ascertain how feasilble it is to port over the functionality
from the deepeval gem to this codebase.

This commit adds the AnswerRelevancy metric. I've focussed on getting the core
functionality in place, and not concerned myself with things like verbose
logs or showing progress bars like deepeval does.

It does the following:

1. passes the answer to the LLM and asks it to break it down into statements
2. passes the question and statements to the LLM and asks it to generte a
   verdict for each statement using the same format as deepeval
3. calculates a score based on the verdicts
4. makes an additional LLM call with the irrelevant statements to get a reason
   for the score.
5. returns a data object with the score and reason

I've added the added the ability to set a threshold and use strict mode
by passing args into the call method. These are optional and default to
0.5 and false respectively.

Were we to implement this properly we would want to add some additional
functionality, but i'm not going to worry about that as part of this spike.
We'd want to add:

- capturing metrics and llm responses
- the ability to configure the model used. I've hardcoded openai.gpt-oss-120b-1:0
  for now.

I've added a basic rake task to generate an answer and evaluate its relevancy.
It follows the pattern of requiring an input env var like our other evaluation tasks.

Anecdotally having run it ~20 times i've not seen any issues that would suggest
we need to implement retries, but were we to add it to our post answer
composition workflow we would use sidekiq to handle this.

I'm going to add it to this workflow is the next commit just to illustrate
how it would integrate.