Abstract Bedrock calls and text content parsing
For each metric we're going to implement, we need to make calls to
Bedrock with the prompt in the user message, parse the response to
get the first text content block and parse the JSON from that.

This new class moves this functionality out of the metric class and into
a dedicated BedrockConverseClient class.