Add BedrockConverseClient and converse stubs
This commit introduces the BedrockConverseClient, a dedicated client for
interacting with AWS Bedrock's Converse API. For now i've kept this
simple since our use case is just to get the answer relevancy metric working.

It allows configuration of the request params via the options, but i suspect
the main one we'll use for now is the messages param since we need to send
the system prompt as the first user message for each LLM call.

I've added a .parse_first_text_content_from_response method which will be
used to extract the text response from the Bedrock Converse API response
as the OpenAI OSS models return reasoning messages in the first content
block.