Add AnswerRelevancy metric
As mentioned in the previous PR we're spiking out adding a deepeval metric
to this codebase to ascertain how feasilble it is to port over the functionality
from the deepeval gem to this codebase.

This adds the answer relevancy metric which is a port of the deepeval
AnswerRelevancy metric.

This commit adds the following classes:

1. AnswerRelevancy::Metric - the main entry point for the metric. This
  class orchestrates the calls to the other classes which handle the
  various LLM calls, scores the verdicts and returns a Result object with
  the score, reason, llm responses and metrics.
2. AnswerRelevancy::StatementGenerator - handles the LLM call which
   generates the statement from the answer.
3. AnswerRelevancy::VerdictsGenerator - handles the LLM call which generates
   the verdicts
4. AnswerRelevancy::ReasonGenerator - handles the LLM call which generates
   the reason for the score.