Add evaluation rake task to run Answer Relevancy Metric
I've added a basic rake task to generate an answer and evaluate its relevancy.
It follows the pattern of requiring an input env var like our other evaluation tasks.

Anecdotally having run it ~20 times i've not seen any issues that would suggest
we need to implement retries, but were we to add it to our post answer
composition workflow we would use sidekiq to handle this.

I'm going to add it to this workflow is the next commit just to illustrate
how it would integrate.