Retry BedrockConverse requests when JSON parsing fails
We can't gaurantee that the OpenAI OSS models will always return valid JSON
responses. It seems to be occur fairly infrequently, but often enough to
warrant a retry mechanism.

This implements a simple retry mechanism that will retry the request up to
3 times if a JSON::ParserError is raised when attempting to parse the response.

We may want to consider adding additional types of errors to retry on in
the future.

Sidekiq will automatically retry the entire job once it's implemented into
the analysis generation workflow, but at the moment, connection errors
etc. are not retried within this class. We might want to add those.