Use Webmock in stubs for Bedrock calls
We use Webmocks to stub out our calls to OpenAI and Bedrock via Anthropic's
SDK. However, in our bedrock stubs we are mocking the client directly.

We're going to be extending our interactions with Bedrock when we integrate
autoevaluation into our application. To keep things consistent and easier to
maintain, we should use Webmock in our Bedrock stubs as well.