Add document types and schemas to search config
Rather than having all the config spread across the separate parser
class files, we want to have it all in one place.

This has a few advantages:

* All the content type -> parser class mappings are in one place, so
  it's easy to find out which parser will be used for a given content
  type
* We have a single place where it's possible to see all the document types
  we're capable of parsing

A rather exhaustive process was followed in this spreadsheet[1] to
identify which document types would be included in indexing. The focus
was based on removing content that would give users misleading
information (such as news that is out of date), irrelevant (such as
deeply technical specialist documents) and documents that contain
personally identifiable information (such as tribunals).

We've included these document types from this document:

* about
* answer
* business_finance_support_scheme
* complaints_procedure
* detailed_guide
* export_health_certificate
* form
* guidance
* guide
* help_page
* html_publication
* international_development_fund
* licence_transaction
* manual
* manual_section
* modern_slavery_statement
* notice
* promotional
* regulation
* service_manual_guide
* simple_smart_answer
* statutory_guidance
* step_by_step_nav
* take_part
* transaction
* travel_advice
* worldwide_organisation

From the spreadsheet, we've not included:

* contact
* guide_part
* how_government_works
* licence
* local_transaction
* place

This differs to how it was in this previous structure, where previously
we stored just the content type -> weight mapping for content types that
deviated from the default weighting of 1.

Now we list _all_ content types we support, but we still just set
weightings on just the ones that deviate from the default. It's assumed
that if a content type is listed here without a weighting, its weighting
is `1`.

We also now store the content types under their respective schema names.
This means we can validate the data we store here against the
data in the `GovukSchemas::Schema` gem. This has a couple of advantages:

* We can't accidentally include a content type that doesn't exist
* We can more strictly validate an incoming document from the message
  queue against its schema and content type. Previously we'd only be
  able to validate the content type.

[1]: https://docs.google.com/spreadsheets/d/1xwgblw2FDKwKVvDXkt33Qe7tc9Wzlqx-wHU3WAEwMi0/edit?gid=0#gid=0