Update parsing logic to use new schema config
Currently each document parser is responsible for defining which schemas
and document types it will parse.

It does that by configuring a map of lambdas which, when passed the
document type, will return a boolean about whether they can be parsed or
not.

We've now moved this config into the `config/search.yml` file instead,
so that it's a bit easier to get an overview of the document types we're
capable of parsing.

This commit updates the logic to check that config file instead of the
constants in each parsing class.