Fix flaky bigquery:export rake spec
This test has been failing on CI for a while. The root cause is that we
create additional answer records when we create the feedback and analysis
in the spec.

The reason this passes the majority of the time despite there being
extra records in the database is due to an idiosyncrasy in the way
freeze_time works.

When you call freeze_time, it sets Time.current to the start of the current
second. Since the majority of the time the answer records are created in the
same second as the freeze_time call, they fall outside the export window
since they're created later in that same second.

This causes the test to fail in the current situation:

1. the 3 answer records are created. We set the created_at of the one
  we explicitly create to 1 hour ago. This record is always exported to
  BigQuery
2. freeze_time is called in the following second which sets Time.current
  to the subsequent second
3. the export task is called, which exports records created between
  2 hours ago and Time.current
4. One or both of the other answer records are returned since they were created
   in the preceeding second, which is within the export window.

We can resolve this issue by associating the feedback and analysis
with the explicitly created answer record, ensuring only one answer
record exists in the database.