Move details and analysis into tabs on question show page
This commit adds the details and analysis tabs to the admin question
show page.

It renders the details/analysis, metrics and llm_responses for each within
their respective tabs.

If no analysis is present, the tabs component doesn't render tabs by default
so we don't have to use it conditionally.