Refactor indexable content item logic
Previously the logic as to whether a parser could parse a content item
was included in that parser's class itself.

This commit moves the logic into the `ContentItemToChunks` class
instead.

The reasoning here is that we can now have this logic in one place so
it's easier to understand. There was also some duplicate logic in some
parser classes.

If we wanted to add more parser classes in the future, we'd need to
duplicate more logic which seems unnecessary.