Rename answer analyses to answer topics
We're moving away from having an answer analyses model to having multiple
smaller models.

The reason for this is that we need to do 3 runs per metric. Due to this
for each metric we're going to have:

- an aggregate model
- a model that represents each run

We still a model to capture answer topics so it makes sense for us to
rename the existing answer analyses model to answer topics.

The one difference is that we're going to be updating llm_responses to
llm_responses since we're only storing the topics response there.

This migration renames the model, adds the llm_response column, backfills
it with the existing llm response and then removes the existing llm_responses
column.