Update all analysis references to topics
This goes through the codebase and updates all references from
"analysis" to "topics" where relevant.

Normally i'd:

- add a new table
- update the code to write to both tables
- backfill data from old table to new table
- switch over to the new table

But since we're currently not live I don't think it's necessary. API
users on integration may have momnentary issues but this should be
minimal.

I've moved the TopicTagger class into the auto_evaluation namespace. I know
technically it's not part of running the metrics, but it will be used to
aggregate auto-evaluation data so i think it makes sense to put it there.

I'll need to ensure I rename the tables on BigQuery before we run
our next export after this PR is merged.