Add schema_name to OpenSearch mapping
We want to start storing the `schema_name` against each document in the
index. This will help us when determining whether to index a particular
piece of content - currently we use the `document_type` to determine
this, but we'd like to validate that against the `schema_name` too when
we pull items off the queue.

The process will be:

1. Add the mapping in this commit
2. Run the `search:update_chunked_content_mappings` Rake task to
   actually add the new mapping to OpenSearch in each env
3. Update the content item syncing code to store the schema name when
   indexing a content item
4. Backfill existing documents in OpenSearch