Add additional bedrock stub to stub all answer relevancy calls
We're going to need to stub out these calls in multiple places so it makes
sense to have a single method that does all the stubbing for us.

I've also prepended stub_ to bedrock_invoke_model_openai_oss_tool_call.
All other stubs have this so it makes sense to be consistent.