Add Rake task for updating schema_name in OpenSearch
This adds a `search:set_schema_name` task that does a bulk update to set
the `schema_name` for documents in the index.

We need this to be set because we want to do a bit more validation on
which document types we process from the queue.

The mapping already exists in OpenSearch for the `schema_name` value to
be set.

Note that this mapping isn't exhaustive. There are some document_types
which we can't automatically infer the schema_name of. The mapping here
contains only instances where there's a 1-1 mapping between
`document_type` and `schema_name`.

We'll need to tackle the rest of the document types that exist in the
index afterwards:

* about
* complaints_procedure
* html_publication
* modern_slavery_statement

This Rake task uses `update_by_query` which is performed server-side in
batches so it's efficient for large numbers of documents[1]. When the
task is run, it only targets documents whose `schema_name` field is
blank, so it can be run again if anything goes wrong and it won't be
loading all documents.

[1]: https://docs.opensearch.org/latest/api-reference/document-apis/update-by-query/