Add schema_name to Result class
Now that we've added the mapping to OpenSearch for `schema_name`, we
need to add the attribute to this `Result` class.

When a document comes out of OpenSearch, all its attributes are used
when instantiating this `Result` class. So if there's an attribute in
the OS document and it's not in this list, it'll raise an
`ArgumentError`.