Implement quota for Answer relevancy metric
This adds the ability to limit the number of answer relevancy metrics
that can be run per hour. If the limit is reached, further evaluations
are skipped and a warning is logged.

We've set the limit to 300 metrics per hour based on discussion with
data science.

A future commits will implement this for topics and add shared examples
for the tests.