Implement quota for auto evaluation
This adds the ability to limit the number of autoevaluations
that can be run per hour. If the limit is reached, further evaluations
are skipped and a warning is logged.

It uses the 'auto_evaluations_count_' prefix combined with the start of
the current hour's timestamp (to_i) to form the cache key. It will expire
after the last evaluation in that hour plus an hour.

When the first evaluation is run in the next hour and new key is created.

We've set the limit to 300 evaluations per hour based on discussion with
data science.

We're treating topic tagging as an auto evaluation. In order to reuse the
quota functionality i've updated the AnswerTopicsJob to ingerit from the
base job and added the quota checks.