Add Coherence metric
This adds the Coherence metric to the auto-evaluation module. Much like
the answer relevancy metric, the Coherence metric uses
BedrockOpenAIOssInvoke to make a tool call the LLM.

It also returns a result object with the same attributes as the answer
relevancy metric result object. I'll move this into a shared data class
in a follow up commit.

There is a difference in score required to normalise the rubric score
to a 0-1 scale. This has been directly ported from our eval codebase
and follows these mappings:

1 - 0.0
2 - 0.25
3 - 0.5
4 - 0.75
5 - 1.0